+++ patches/src/main/kotlin/app/revanced/patches/instagram/misc/devmenu/Fingerprints.kt
@@ -1,0 +1,12 @@
+
+package app.revanced.patches.instagram.misc.devmenu
+
+import app.revanced.patcher.fingerprint
+
+internal val clearNotificationReceiverFingerprint by fingerprint {
+    custom { method, classDef ->
+        method.name == "onReceive" &&
+            classDef.type == "Lcom/instagram/notifications/push/ClearNotificationReceiver;"
+    }
+    strings("NOTIFICATION_DISMISSED")
+}