+++ patches/src/main/kotlin/app/revanced/patches/spotify/layout/hide/createbutton/HideCreateButtonPatch.kt
@@ -17,7 +17,8 @@ private const val EXTENSION_CLASS_DESCRIPTOR =
 @Suppress("unused")
 val hideCreateButtonPatch = bytecodePatch(
     name = "Hide Create button",
-    description = "Hides the \"Create\" button in the navigation bar."
+    description = "Hides the \"Create\" button in the navigation bar. The latest app targets do not need this patch.",
+    use = false
 ) {
     compatibleWith("com.spotify.music")